Call Custom Probes and fix Infrastructure Typo
