+++ src/AspNetCore.AutoHealthCheck.Tests/Infrastructure/ResultProcessor/HealtCheckResultProcessorTests.cs
@@ -25,16 +25,34 @@ using System.Diagnostics;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
 using Moq;
 using Xunit;
 
-namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
+namespace AspNetCore.AutoHealthCheck.Tests.Infrastructure.ResultProcessor
 {
-    public class HealtCheckResultProcessorNegativeTests
+    public class HealtCheckResultProcessorTests
     {
         [Fact]
-        public async Task HealtCheckResultProcessor_should_fail_returning_500_with_default_rule()
+        public async Task HealtCheckResultProcessor_should_return_ok_with_emtpty_results()
+        {
+            // arrange
+            var watch = new Stopwatch();
+            watch.Start();
+            var contex = new Mock<IAutoHealthCheckContext>();
+            contex.Setup(c => c.Configurations)
+                .Returns(new AutoHealthCheckConfigurations());
+
+            // act
+            var result = await HealtCheckResultProcessor.ProcessResult(contex.Object, watch, new HttpResponseMessage[0]);
+
+            // assert
+            Assert.NotNull(result);
+            Assert.Equal(200, (int)result.HttpStatus);
+            Assert.True(result.Success);
+        }
+
+        [Fact]
+        public async Task HealtCheckResultProcessor_should_return_ok_with_default_rule()
         {
             // arrange
             var watch = new Stopwatch();
@@ -46,8 +64,10 @@ namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
 
             var messages = new List<HttpResponseMessage>
             {
-                new HttpResponseMessage(HttpStatusCode.InternalServerError),
-                new HttpResponseMessage(HttpStatusCode.BadGateway)
+                new HttpResponseMessage(HttpStatusCode.OK),
+                new HttpResponseMessage(HttpStatusCode.AlreadyReported),
+                new HttpResponseMessage(HttpStatusCode.Ambiguous),
+                new HttpResponseMessage(HttpStatusCode.BadRequest)
             };
 
             // act
@@ -55,15 +75,12 @@ namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
 
             // assert
             Assert.NotNull(result);
-            Assert.Equal(500, (int)result.HttpStatus);
-            Assert.False(result.Success);
-            Assert.Equal(2, result.UnhealthyEndpoints.Count);
-            Assert.Equal(500, result.UnhealthyEndpoints[0].HttpStatusCode);
-            Assert.Equal(502, result.UnhealthyEndpoints[1].HttpStatusCode);
+            Assert.Equal(200, (int)result.HttpStatus);
+            Assert.True(result.Success);
         }
 
         [Fact]
-        public async Task HealtCheckResultProcessor_should_fail_with_custom_rule()
+        public async Task HealtCheckResultProcessor_should_return_ok_with_custom_rule()
         {
             // arrange
             var watch = new Stopwatch();
@@ -79,7 +96,9 @@ namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
             var messages = new List<HttpResponseMessage>
             {
                 new HttpResponseMessage(HttpStatusCode.AlreadyReported),
-                new HttpResponseMessage(HttpStatusCode.OK)
+                new HttpResponseMessage(HttpStatusCode.Ambiguous),
+                new HttpResponseMessage(HttpStatusCode.BadRequest),
+                new HttpResponseMessage(HttpStatusCode.InternalServerError),
             };
 
             // act
@@ -87,13 +106,12 @@ namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
 
             // assert
             Assert.NotNull(result);
-            Assert.Equal(500, (int) result.HttpStatus);
-            Assert.False(result.Success);
-            Assert.Single(result.UnhealthyEndpoints);
+            Assert.Equal(200, (int)result.HttpStatus);
+            Assert.True(result.Success);
         }
 
         [Fact]
-        public async Task HealtCheckResultProcessor_should_fail_returning_custom_response_code_with_default_rule()
+        public async Task HealtCheckResultProcessor_should_return_custom_response_code_with_default_rule()
         {
             // arrange
             var watch = new Stopwatch();
@@ -103,13 +121,15 @@ namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
             contex.Setup(c => c.Configurations)
                 .Returns(new AutoHealthCheckConfigurations
                 {
-                    DefaultUnHealthyResponseCode = HttpStatusCode.AlreadyReported
+                    DefaultHealthyResponseCode = HttpStatusCode.Continue
                 });
 
             var messages = new List<HttpResponseMessage>
             {
-                new HttpResponseMessage(HttpStatusCode.InternalServerError),
-                new HttpResponseMessage(HttpStatusCode.BadGateway)
+                new HttpResponseMessage(HttpStatusCode.OK),
+                new HttpResponseMessage(HttpStatusCode.AlreadyReported),
+                new HttpResponseMessage(HttpStatusCode.Ambiguous),
+                new HttpResponseMessage(HttpStatusCode.BadRequest)
             };
 
             // act
@@ -117,11 +137,8 @@ namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure.ResultProcessor
 
             // assert
             Assert.NotNull(result);
-            Assert.Equal(HttpStatusCode.AlreadyReported, result.HttpStatus);
-            Assert.False(result.Success);
-            Assert.Equal(2, result.UnhealthyEndpoints.Count);
-            Assert.Equal(500, result.UnhealthyEndpoints[0].HttpStatusCode);
-            Assert.Equal(502, result.UnhealthyEndpoints[1].HttpStatusCode);
+            Assert.Equal(HttpStatusCode.Continue, result.HttpStatus);
+            Assert.True(result.Success);
         }
     }
 }
