+++ src/AspNetCore.AutoHealthCheck.Tests/Infrastructure/Scapers/RoutingScraperTests.cs
@@ -26,7 +26,7 @@ using AspNetCore.AutoHealthCheck.Tests.Controllers;
 using Microsoft.AspNetCore.Mvc.Controllers;
 using Xunit;
 
-namespace AspNetCore.AutoHealthCheck.Tests.Infraestructure
+namespace AspNetCore.AutoHealthCheck.Tests.Infrastructure.Scapers
 {
     public class RoutingScraperTests
     {
