+++ src/AspNetCore.AutoHealthCheck/AutoHealthCheckServiceCollectionExtensions.cs
@@ -30,6 +30,8 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class AutoHealthCheckServiceCollectionExtensions
     {
+        private static AutoHealthCheckContextAccessor _contextAccessor;
+
         /// <summary>
         ///     Add Auto health check to the asp.net core application without configurations
         /// </summary>
@@ -53,13 +55,35 @@ namespace Microsoft.Extensions.DependencyInjection
             // resolve options
             var options = new AutoHealthCheckConfigurations();
             setupAction?.Invoke(options);
-            var accessor = new AutoHealthCheckContextAccessor();
-            accessor.SetConfigurations(options);
-            services.AddSingleton<IAutoHealthCheckContextAccessor>(accessor);
+            _contextAccessor = new AutoHealthCheckContextAccessor();
+            _contextAccessor.SetConfigurations(options);
+            services.AddSingleton<IAutoHealthCheckContextAccessor>(_contextAccessor);
 
             // check if the service need to run automatically
             if (options.AutomaticRunConfigurations.AutomaticRunEnabled)
                 services.AddSingleton<IHostedService, AutoHealtCheckProcess>();
+
+            return services;
+        }
+
+        /// <summary>
+        ///     Add custom probe work.
+        /// </summary>
+        /// <typeparam name="TProbe">Probe type to add to the work queue.</typeparam>
+        /// <param name="services"></param>
+        /// <returns></returns>
+        public static IServiceCollection AddCustomProbe<TProbe>(this IServiceCollection services)
+            where TProbe : class, IProbe
+        {   
+            if (_contextAccessor == null)
+                throw new InvalidOperationException("Please first call AddAutoHealthCheck method.");
+
+            // register probe in the IOC engine.
+            services.AddTransient(typeof(TProbe));
+
+            // add the probe to context
+            var context = (AutoHealthCheckContext)_contextAccessor.Context;
+            context.AddProbe<TProbe>();
 
             return services;
         }
