+++ src/AspNetCore.AutoHealthCheck/Infrastructure/AsyncLazy.cs
