+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Context/AutoHealthCheckContextAccessor.cs
