+++ src/AspNetCore.AutoHealthCheck/Infrastructure/CustomProbes/IProbe.cs
