+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Discovery/InternalRouteInformationEvaluator.cs
