+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Discovery/RouteDiscover.cs
