+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Discovery/Scrapers/BodyContentScraper.cs
