+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Discovery/Scrapers/QueryStringScraper.cs
