+++ src/AspNetCore.AutoHealthCheck/Infrastructure/EndpointCaller.cs
