+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Endpoints/Endpoint.cs
