+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Endpoints/EndpointBuilder.cs
