+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Endpoints/EndpointMessageTranslator.cs
