+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Endpoints/HealthyResponse.cs
