+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Endpoints/IEndpoint.cs
