+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Endpoints/UnhealthyEndpoint.cs
