+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Extensions/RouteInformationExtensions.cs
