+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Extensions/TypeExtensions.cs
