+++ src/AspNetCore.AutoHealthCheck/Infrastructure/HealtCheckResultProcessor.cs
