+++ src/AspNetCore.AutoHealthCheck/Infrastructure/HostedServices/AutoHealtCheckProcess.cs
