+++ src/AspNetCore.AutoHealthCheck/Infrastructure/IEndpointCaller.cs
