+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Middlewares/AutoHealtCheckMiddlewareOptions.cs
