+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Middlewares/AutoHealthCheckMiddleware.cs
