+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Plugins/IHealtCheckResultPlugin.cs
