+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Plugins/IHttpEndpointPlugin.cs
