+++ src/AspNetCore.AutoHealthCheck/Infrastructure/Plugins/IPlugin.cs
