+++ src/WebApplication/Probes/CustomProbe.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading.Tasks;
+using AspNetCore.AutoHealthCheck;
+
+namespace WebApplication.Probes
+{
+    public class CustomProbe : IProbe
+    {
+        public string Name => typeof(CustomProbe).Name;
+
+        public Task<bool> Check()
+        {
+            return Task.FromResult(false);
+        }
+    }
+}
