+++ src/WebApplication/Startup.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Net;
+using AspNetCore.AutoHealthCheck;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using WebApplication.Plugins;
+using WebApplication.Probes;
 
 namespace WebApplication
 {
@@ -29,7 +31,8 @@ namespace WebApplication
                 c.AutomaticRunConfigurations.BaseUrl = new Uri("http://localhost:50387");
                 c.AutomaticRunConfigurations.SecondsInterval = 1;
                 c.ResultPlugins.Add(new ResultPlugin());
-            });
+            })
+            .AddCustomProbe<CustomProbe>();
         }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
@@ -38,7 +41,7 @@ namespace WebApplication
             app.UseMvc();
             app.UseAutoHealthCheck(c =>
             {
-                c.RoutePrefix = "insights/healtcheck";
+                c.RoutePrefix = "insights/healthcheck";
             });
         }
     }
